+++ src/student_jekaterina_soldatova/lesson_7/level_1_intern/WordService.java
@@ -3,12 +3,6 @@ package student_jekaterina_soldatova.lesson_7.level_1_intern;
 import java.util.Scanner;
 
 class WordService {
-    public static void main(String[] args) {
-        String sentence = WordService.getString(); //get string
-        String[] arrayFromString = WordService.stringToArray(sentence); //string to array
-        String mostFrequentWord = WordService.findMostFrequentWord(arrayFromString); //find most frequent word
-        WordService.printResult(mostFrequentWord); //print results
-    }
     static String getString() {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Please enter a sentence");
